Extract ParsePipe to parse.ts
