+++ lib/syntax/parse.d.ts
@@ -1,0 +1,4 @@
+import { Context, Frame } from "../frames";
+export declare class ParsePipe extends Frame {
+    constructor(out: Frame, meta?: Context);
+}
