+++ lib/syntax/parse.js
@@ -1,0 +1,19 @@
+"use strict";
+var __extends = (this && this.__extends) || function (d, b) {
+    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+    function __() { this.constructor = d; }
+    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+};
+var frames_1 = require("../frames");
+var ParsePipe = (function (_super) {
+    __extends(ParsePipe, _super);
+    function ParsePipe(out, meta) {
+        if (meta === void 0) { meta = frames_1.Void; }
+        var _this = this;
+        meta[frames_1.Frame.kOUT] = out;
+        _this = _super.call(this, meta) || this;
+        return _this;
+    }
+    return ParsePipe;
+}(frames_1.Frame));
+exports.ParsePipe = ParsePipe;
