+++ lib/syntax/pipeline.js
@@ -6,6 +6,7 @@ var __extends = (this && this.__extends) || function (d, b) {
 };
 var frames_1 = require("../frames");
 var lex_1 = require("./lex");
+var parse_1 = require("./parse");
 var EvalPipe = (function (_super) {
     __extends(EvalPipe, _super);
     function EvalPipe(out, meta) {
@@ -18,23 +19,11 @@ var EvalPipe = (function (_super) {
     return EvalPipe;
 }(frames_1.Frame));
 exports.EvalPipe = EvalPipe;
-var ParsePipe = (function (_super) {
-    __extends(ParsePipe, _super);
-    function ParsePipe(out, meta) {
-        if (meta === void 0) { meta = frames_1.Void; }
-        var _this = this;
-        meta[frames_1.Frame.kOUT] = out;
-        _this = _super.call(this, meta) || this;
-        return _this;
-    }
-    return ParsePipe;
-}(frames_1.Frame));
-exports.ParsePipe = ParsePipe;
 var piper = function (input, context) {
     if (context === void 0) { context = frames_1.Void; }
     var result = new frames_1.FrameArray([], context);
     var evaluator = new EvalPipe(result);
-    var parser = new ParsePipe(evaluator);
+    var parser = new parse_1.ParsePipe(evaluator);
     var lexer = new lex_1.LexPipe(parser);
     var status = lexer.lex_string(input);
     if (status !== lexer) {
@@ -52,7 +41,7 @@ exports.framify = function (input, context) {
 exports.framify_new = function (input, context) {
     if (context === void 0) { context = frames_1.Void; }
     var result = new frames_1.FrameArray([], context);
-    var parser = new ParsePipe(result);
+    var parser = new parse_1.ParsePipe(result);
     var status = pipe(input, parser);
     console.error("\n* framify_new.pipe returned " + status);
     var expr = result.at(0);
