+++ src/syntax/parse.ts
@@ -1,0 +1,8 @@
+import { Context, Frame, FrameString, FrameSymbol, Void } from "../frames";
+
+export class ParsePipe extends Frame {
+  constructor(out: Frame, meta: Context = Void) {
+    meta[Frame.kOUT] = out;
+    super(meta);
+  }
+}
