+++ src/syntax/pipeline.ts
@@ -1,14 +1,8 @@
 import { Context, Frame, FrameArray, FrameLazy, FrameString, FrameSymbol, Void } from "../frames";
 import { LexPipe } from "./lex";
+import { ParsePipe } from "./parse";
 
 export class EvalPipe extends Frame {
-  constructor(out: Frame, meta: Context = Void) {
-    meta[Frame.kOUT] = out;
-    super(meta);
-  }
-}
-
-export class ParsePipe extends Frame {
   constructor(out: Frame, meta: Context = Void) {
     meta[Frame.kOUT] = out;
     super(meta);
